Various improvements to QOL, code documentation and comprehensibility
Includes, but is not limited to, the following changes: - Added blur and opacity to polish the start selection screen. - Rewrote some parts of the code to be more comprehensible and easy to modify. - Added documentation to everything I could. - Fixed targets not updating after deleting and re-adding the artillery system.


Former-commit-id: f90f98bf3e714932fdc9acc2aded4f342d8f0e6f
Former-commit-id: 3ea78073fe54b2962745d1b44ce797794645503d