Various improvements to QOL, code documentation and comprehensibility
Includes, but is not limited to, the following changes: - Added blur and opacity to polish the start selection screen. - Rewrote some parts of the code to be more comprehensible and easy to modify. - Added documentation to everything I could. - Fixed targets not updating after deleting and re-adding the artillery system.